Initial commit
This implements encoding a file to a PDF, and restoring the file from a
set of images.  There is no encryption; however, you need some number of
QR codes in order to restore the file, so it is possible to divide the
resulting pages to avoid premature decoding.

Currently this using `rxing` to decode the images, but `qrcode` to
encode them.  It is hoped that we can switch to `rxing` for both at some
future point.

There is an integration test, in which we encode then decode the license
file and check that it survived the round trip.  Note that this seems to
indicate that even with perfect images, we will fail to scan some QR
codes.